+++ jobscripts/templates/slurm_pipeline_S6c-6d.sh
@@ -3,7 +3,7 @@
 # Usage: sbatch slurm_pipeline_S6c-6d.sh
 
 # IMPORTANT: before running this step, use the output from Step 6b to update the effective genome
-#            size value in the file "config/chromatin_assembly_config.yml", in Section 4: Intermediate output
+#            size value in the file "config/chromatin_assembly_config.yml", in Section 1: Input data and parameters
 
 # NOTE: Depending on your data, you may need to update the resources in this job script or
 #       in the Snakemake slurm profile (chromatin-assembly/profiles/slurm/config.yaml)
@@ -16,8 +16,8 @@
 #SBATCH --mail-user=caitlin.cherryh@csiro.au
 #SBATCH --mail-type=END,FAIL,TIME_LIMIT
 #SBATCH --account=OD-233464
-#SBATCH --error=/scratch3/che318/chromatin-assembly/log/slurm_%j_%x.err
-#SBATCH --out=/scratch3/che318/chromatin-assembly/log/slurm_%j_%x.out
+#SBATCH --error=/scratch3/che318/chromatin-assembly/logs/slurm_%j_%x.err
+#SBATCH --out=/scratch3/che318/chromatin-assembly/logs/slurm_%j_%x.out
 
 # ----------------Modules------------------------- #
 module load python