+++ CHANGELOG.md
@@ -6,7 +6,27 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
   
 ## [Unreleased] - yyyy-mm-dd
+
+Improved read merging and extended run time for long steps.
+
+### Added
+- Better template data in 'data/template/`
+
+### Changed
+- S0 read merging - BASH script with simpler input parameters and improved 
+internal logic 
+- Moved vars for effective genome size and species binomial into Section 1 of 
+the config file
+- Improved master `README.md` file
+
+### Fixed
+- Specify `ext` partition for long steps (S3/S4), allowing run time of up to 30 
+days
  
+## [0.1.1] - 2025-05-15
+
+Multi-sample input now available.
+
 ### Added
 - R script (`workflow/R/s0_merge_reads.R`) to merge reads into one R1 and one R2 
 file per sample name 
@@ -56,8 +76,7 @@ and `profiles/slurm/config.yaml`
   - Removed hard coded variables and reference from config file every time instead
   - Removed absolute paths and replaced with paths relative to the working directory
   `chromatin-assembly/`
- 
- 
+
 ## [0.1.0] - 2025-04-08
 
 First release. Both minor and major improvements still pending.